requirements.txt
pyfiglet==1.0.2
google-generativeai==0.6.0
unstructured[all-docs]
unstructured[local-inference]
libmagic==1.0
pytesseract==0.3.10
faiss-cpu==1.8.0.post1
langchain==0.0.336
sentence-transformers==3.0.1
python-dotenv==1.0.0